requirements.txt
SpeechRecognition
pyttsx3
pywhatkit
pyautogui
wikipedia
pygame
requests
opencv-python
geopy
gTTS
google-genai
winsdk
